Fixed initial HEADERS loading problem
