+++ src/constants/paths.js
@@ -1,0 +1,1 @@
+export const CONFIG_PATH = '/etc/sundial/config.json';
